+++ Sources/PortalSwift/Utils/Blockchains/PortalBlockchain.swift
@@ -10,12 +10,14 @@ import Foundation
 public class PortalBlockchain {
   static let mainnetReferences: [PortalNamespace: String] = [
     .eip155: "1",
-    .solana: "5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp"
+    .solana: "5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp",
+    .xrpl: "0"
   ]
   static let namespaceCurves: [PortalNamespace: PortalCurve] = [
     .eip155: .SECP256K1,
     .bip122: .SECP256K1,
-    .solana: .ED25519
+    .solana: .ED25519,
+    .xrpl: .SECP256K1
   ]
   static let namespaceSignerMethods: [PortalNamespace: [PortalRequestMethod]] = [
     .eip155: [